[add] 모바일 버전 알람 추가
