+++ src/components/notification/NotificationList.tsx
@@ -4,12 +4,28 @@ import {Grid} from "../../../assets/style/Layout.style";
 import {MarkdownBase, MarkdownSm} from "../../../assets/style/Markdown.style";
 import {Color} from "../../../assets/style/Color.style";
 import styled from "@emotion/styled";
+import {media} from "../../../assets/style/Media.style";
 
 const NotiList = styled.li`
   padding:1.5em; 
   cursor:pointer; 
   ${Grid(25, 75)}; 
   ${MarkdownSm()}
+  ${media.sm`padding:1.5em 0;`}
+`
+
+interface notiTypeProps{
+    isRead:string | any;
+}
+const NotiType=styled.div<notiTypeProps>`
+  margin-right:0.8em; 
+  text-align:center; 
+  padding:0.3em;  
+  border-radius:0.5em;
+  ${(props:notiTypeProps)=>props.isRead ? 
+    css`${MarkdownSm(Color.purple100)}; border:1px solid ${Color.purple100};`: 
+    css`${MarkdownSm(Color.purple200)}; border:1px solid ${Color.purple200};`
+    };
 `
 
 interface propsType {
@@ -26,11 +42,12 @@ const NotificationList: React.FC<propsType> = ({notificationList, readOne}) => {
                     notificationList.map((item: any, index: number) => {
                         return (
                             <NotiList key={`noti-${index}`}
-                                      css={css`${item.read_at ? `` : `background-color:${Color.purple100}`}`}
+                                      css={css`${item.read_at ? `color:${Color.gray200}`: ''}`}
                                       onClick={() => readOne(item.id, index)}>
                                 <div>
-                                    <div
-                                        css={css`margin-right:0.8em; text-align:center; padding:0.3em; ${MarkdownSm(Color.purple200)}; border:1px solid ${Color.purple200}; border-radius:0.5em;`}>{item.notice_type.description}</div>
+                                    <NotiType isRead={item.read_at}>
+                                        {item.notice_type.description}
+                                    </NotiType>
                                 </div>
                                 <div>
                                     <p css={css`margin-bottom:0.3em;`}>{item.contents}</p>
